+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## UNRELEASED
 
+- Added reference to the state manager to InitStateContext and UpdateStateContext.
 - Added "find_ancestor(ancestor_type)" method to StateAddress and CallContext.
 
 ## 1.7.0 (2023-12-13)